Clean up conductivity calculations.
